F/GO : Remove all support
Someone could do this better than I did, and people actually did.
It's sad to see that the whole reason this bot was born is now gone, but I think it's better to leave this field in the hands of better people.

First commit in Lunar New Year 2022 is a commit that removes stuff, and I wonder if this is good :thonk: